Remove arch-specific parts of libmath
These are just performance optimizations and also currently using just the obsolete slow x87 instructions on x86 platforms. These also hide the brokenness of the soft-implementations on the most used platforms. In principle, complicating the build with optimizations of a broken library doesn't make sense.
