--- uspace/lib/math/arch/ia32/include/libarch/x87.h
@@ -1,1 +1,0 @@
-../../../amd64/include/libarch/x87.h
